Split denoise implementation to minimize dependencies
The main goal here is to have denoise require the fewest possible dependencies when running under root, which avoids issues with PYTHONPATH, dependencies being installed by the user, and generally to minimize the capabilities of denoise.

- denoise_client now provides the interface for rebench to access denoise
- minor refactorings to avoid circular dependencies
- move getting number of cores to denoise_client to avoid dependency on cpuinfo inside root
- cache number of cpu cores to avoid requesting them repeatedly which is very expensive
